Crate fogcore

A library for rendering a person’s living tracks.

The library current supports importing data from the Fog of World App, and render it into images using image library.

Usage

Please refer to the examples folder.

Re-exports

pub use renderer::TileRendererBasic;

pub use renderer::TileRendererPremium;

pub use renderer::TileRendererPremium2;

pub use renderer::TileRendererTrait;

pub use renderer::TileShader;

pub use fogmaps::FogMap;

Modules

fogmaps

renderer

Macros

log_print

Enums

TileSize

Constants

DEFAULT_BG_COLOR2

DEFAULT_FG_COLOR2

DEFAULT_TILE_SIZE

DEFAULT_VIEW_SIZE_POWER

Functions

image_to_png_data

lat_to_tile_y

lng_to_tile_x

load_tracks_map_folder

set_panic_hook

tile_x_y_to_lng_lat