Struct fogcore::fogmaps::FogMap

pub struct FogMap {
    pub tiles: HashMap<(i64, i64), Tile>,
}

An in-memory efficient representation of a persons tracks on the Earth.

Fields
tiles: HashMap<(i64, i64), Tile>
Implementations

impl FogMap

pub fn new() -> Self

Creates an empty FogMap

pub fn add_fow_file(&mut self, file_name: &str, data: Vec<u8>)

Adds tracks by importing from a data file of the Fog of World App.

Note that this operations will NOT REPLACE the existing tracks in FogMap, this operation is purely incremental.

pub fn get_tile(&self, x: i64, y: i64) -> Option<&Tile>

pub fn lng_lat_to_tile_x_y(lng: f64, lat: f64, zoom: i16) -> (i64, i64)

pub fn add_line( &mut self, start_lng: f64, start_lat: f64, end_lng: f64, end_lat: f64, )